+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 
 1.2.1 (2026-06-02)
 ------------------
-* Added swerve odometry and `/odom` publishing for SH5 DDS bringup.
-* Tuned FFW_SH5 XM350 actuator parameters.
+* Added swerve odometry calculation. 
+* Updated `/odom` publishing in SH5 DDS bringup.
 * Contributors: Howon Kim
 
 1.2.0 (2026-05-28)